+++ particlegan/__init__.py
@@ -11,7 +11,8 @@ from .grad_regularizers import GradientPenalty
 from .k3p import CriticAnchor, CriticSpikeGuard, DirectParticleResponse, LatentRowDamping
 from .locked_shared import LOCKED_SHARED, locked_adv_defaults, make_b_cap, make_gan_loss
 from .particle_prior import GaussianPrior, MoGParticlePrior, ParticlePrior, calibrate_mog_sigma
-from .recipes import Recipe, get_recipe, learning_rate_scale, learning_rate_scales
+from .recipes import (Recipe, get_recipe, learning_rate_scale, learning_rate_scales,
+                      scale_learning_rates)
 from .training import GANTrainer, K3PCritic
 from .vicreg_loss import ParticleRegularizer
 
@@ -20,7 +21,7 @@ __all__ = [
     "ParticlePrior", "MoGParticlePrior", "calibrate_mog_sigma", "GaussianPrior", "GANLoss", "GradientPenalty",
     "CriticAnchor", "CriticSpikeGuard", "LatentRowDamping", "DirectParticleResponse",
     "ParticleRegularizer", "DDGAN", "UCD", "ucd_labels", "ucd_loss", "ucd_scores",
-    "Recipe", "get_recipe", "learning_rate_scale", "learning_rate_scales", "GANTrainer", "K3PCritic",
+    "Recipe", "get_recipe", "learning_rate_scale", "learning_rate_scales", "scale_learning_rates", "GANTrainer", "K3PCritic",
     "BatchDistanceDiscriminator", "LinearSkipDiscriminator",
     "LOCKED_SHARED", "locked_adv_defaults", "make_gan_loss", "make_b_cap",
 ]